Added a test suite
I added some unittests for some rules from EBAE '94. That is, up to and
including Rule I. ;-) These tests demonstrate each part of the rule, and
so largely consist of [fair] edge-cases I had [largely] not considered.
All-in-all, however, I fixed more errors in the tests than in the
program. Still, the flurry of recent bugfixes was all due to running
(and) verifying the tests themselves.

There are three types of tests from the book, ordered by fun,
descending:
- Those that still fail
- Those that now succeed
- Those that no known deterministic algorithm can complete in a
  reasonable amount of time.

Of course, the program currently passes all tests, which means I haven't
implemented all of them that I could have: There are a few nagging ones
that fall into the first category.

An example of a rule that the current algorithm will always fail (3rd
category) is the use of an apostrophe in the string "EBAE '94". This
single-quote /should/ be interpreted as an apostrophe, but that relies
on semantic info I introduced earlier, i.e. you and I know that 94 is a
year. Of course, one day we can implement an uber-sketch regex heuristic
that is wrong less often than always.