Created an argparser factory method
argparseFactory() receives as input a specification of the
argparse.ArgumentParser that you want to create, including any group
definitions. It automagically generates all of the commands for you,
with less effort on your part. This is useful, so maybe I should patch
argparse itself with this sometime.

It does NOT support sub-commands, yet, but I don't need them yet. :-p